\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap loop}
\usage{
  bootstrap(X, B = 1000, dt = 0.1)
}
\description{
  Bootstrapping method, using the Johansen procedure for estimation of statistic-distribution (only for p=3 at this moment...)
}
\arguments{
  \item{X}{Data matrix}
  \item{B}{Number of iterations}
  \item{dt}{}
}
\examples{
  hello()
}
