\name{hilbert}
\alias{hilbert}
\title{Performs Hilbert transform}
\usage{
hilbert(inVec, n = length(inVec))
}
\arguments{
\item{inVec}{X coordinates of the oscillating process}
}
\description{
Performs Hilbert transform and returns analytic signal
}
\value{
Complex vector. The real part are X-coordinates, imaginary part are Y-coordinates of the oscillator.
}
\examples{
hello()
}
